Added an auto restart and winner inidaction
Set the text at the bottom to declare a winner and also a way for the game to auto restart.

The game will remove all functionality if somebody wins using the inert attribute, however, we have to remove the attribute when recreating the game board for another round.

(Inert attribute is not currently supported by FireFox as of 26/03/2023 yet)